move join action into new class `JoinActions`
